requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
langchain-openai
langchain
openai
youtube-transcript-api
requests
pyyaml
tqdm